feat: QuestionDetailNav 공통화 및 isSharable 속성 추가
