+++ src/_components/QuestionDetail/questionDetail.css.ts
@@ -26,17 +26,3 @@ export const categoryHighlight = style({
   // color: primary,
   fontWeight: 600,
 });
-
-export const shareButton = style({
-  color: 'white',
-  fontSize: '1.4rem',
-  backgroundColor: '#00A86B',
-  padding: '1rem 2rem',
-  borderRadius: '10rem',
-  fontWeight: 600,
-
-  transition: 'transform 0.4s ease-in-out',
-  ':hover': {
-    transform: 'scale(1.03)',
-  },
-});
